viewAddDevice: - merge, fix

+++ app/src/main/java/com/example/hackathon2022/domain/DomainHome.kt
@@ -3,6 +3,6 @@ package com.example.hackathon2022.domain
 data class DomainHome(
     var id: Int? = null,
     var homeName: String = "",
-    var usagePower: Int? = null,
+    var usagePower: Double? = null,
     var devices: MutableList<DomainDevice> = mutableListOf(),
 )
